requirements.txt
Pyyaml
PySimpleGUI
PyInstaller
spoonacular